Added anti troll mechanism
When the patient makes 5 or more change requests or creates an
appointment 8 or more times in the last 30 days, the system blocks the
user from the app. The user will no longer be able to login to the app.
Also changed the AppointmentChangeRequest attribute NewRequestType name
to RequestType. Added AppointmentChangeRequest loading from
PatientDataManager.